requirements.txt

streamlit
google-generativeai
speechrecognition
pyttsx3
pillow
requests
gtts
pyaudio
#pip install streamlit google-generativeai SpeechRecognition pyttsx3 pyaudio gTTS pipwin python-dotenv
